# src/main/resources/static/api-docs.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: CTA Manager
  description: API para operações no CTA
servers:
  - url: http://localhost:9090/api/v1/
    description: local
  - url: https://cta-{environment}.ssp.ma.gov.br/api/v1/
    variables:
      environment:
        default: api.dev
        enum:
          - api
          - api.dev
          - api.staging
tags:
  - name: Cadastro
paths:
  /fabricante:
    get:
      summary: Buscar todos os fabricantes de aeronave
      operationId: buscarFabricantes
      tags:
        - Aeronave
      responses:
        200:
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fabricante'
    post:
      summary: Salvar um fabricante de aeronave
      operationId: adicionarFabricante
      tags:
        - Aeronave
      requestBody:
        $ref: '#/components/requestBodies/FabricanteBody'
      responses:
        200:
          description: Criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fabricante'
  /fabricante/{id}:
    parameters:
      - name: id
        in: path
        description: ID do Fabricante
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Aeronave
      summary: Buscar um fabricante de aeronave por Id
      operationId: buscarFabricantePorId
      responses:
        200:
          description: Fabricante encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fabricante'
        404:
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Remover um fabricante de aeronave por Id
      operationId: removerFabricante
      tags:
        - Aeronave
      responses:
        204:
          description: Fabricante removido com sucesso
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Atualizar um fabricante de aeronave
      operationId: atualizarFabricante
      tags:
        - Aeronave
      requestBody:
        $ref: '#/components/requestBodies/FabricanteBody'
      responses:
        200:
          description: Criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fabricante'

        404:
          $ref: '#/components/responses/NotFound'
  /aeronave:
    get:
      summary: Buscar todas as aeronaves
      operationId: buscarAeronaves
      tags:
        - Aeronave
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aeronave'
    post:
      summary: Salvar aeronave
      operationId: adicionarAeronave
      requestBody:
        $ref: '#/components/requestBodies/AeronaveBody'
      tags:
        - Aeronave
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aeronave'
  /aeronave/{id}:
    parameters:
      - name: id
        in: path
        description: ID da aeronave
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Aeronave
      summary: Buscar uma aeronave por Id
      operationId: buscarAeronavePorId
      responses:
        200:
          description: Aeronave encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aeronave'
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Atualizar uma aeronave
      operationId: atualizarAeronave
      tags:
        - Aeronave
      requestBody:
        $ref: '#/components/requestBodies/AeronaveBody'
      responses:
        200:
          description: Criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aeronave'

        404:
          $ref: '#/components/responses/NotFound'
  /aeronave/{id}/horimetro:
    parameters:
      - name: id
        in: path
        description: ID da aeronave
        required: true
        schema:
          type: integer
          format: int64
    post:
      summary: Adicionar horas de voo e manutenção
      operationId: adicionarHoras
      tags:
        - Aeronave
      requestBody:
        $ref: '#/components/requestBodies/HorimetroBody'
      responses:
        200:
          description: Horimetro adicionado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Horimetro'
        404:
          $ref: '#/components/responses/NotFound'
  /pessoa:
    get:
      summary: Buscar todas as pessoas
      operationId: buscarPessoas
      tags:
        - Cadastro
      parameters:
        - in: query
          name: nome
          schema:
            type: string
          description: Buscar pessoas pelo nome
        - in: query
          name: cpf
          schema:
            type: string
          description: Buscar pessoa pelo cpf
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pessoa'
    post:
      summary: Adicionar uma pessoa
      operationId: adicionarPessoa
      requestBody:
        $ref: '#/components/requestBodies/PessoaBody'
      tags:
        - Cadastro
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pessoa'
  /pessoa/{id}:
    parameters:
    - name: id
      in: path
      description: ID da aeronave
      required: true
      schema:
        type: integer
        format: int64
    get:
      tags:
        - Cadastro
      summary: Buscar uma pessoa por Id
      operationId: buscarPessoaPorId
      responses:
        200:
          description: Pessoa encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pessoa'
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Atualizar uma pessoa
      operationId: atualizarPessoa
      tags:
        - Cadastro
      requestBody:
        $ref: '#/components/requestBodies/PessoaBody'
      responses:
        200:
          description: Criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pessoa'
        404:
          $ref: '#/components/responses/NotFound'
  /aeronauta:
    get:
      summary: Buscar todos os aeronautas
      operationId: buscarAeronauta
      tags:
        - Cadastro
      parameters:
        - in: query
          name: codigo_anac
          schema:
            type: integer
          description: Buscar aeronauta pelo codigo anac
        - in: query
          name: cpf
          schema:
            type: string
          description: Buscar aeronauta pelo cpf
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aeronauta'
    post:
      summary: Adicionar um aeronauta
      operationId: adicionarAeronauta
      requestBody:
        $ref: '#/components/requestBodies/AeronautaBody'
      tags:
        - Cadastro
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aeronauta'
  /aeronauta/{id}:
    parameters:
    - name: id
      in: path
      description: ID da aeronave
      required: true
      schema:
        type: integer
        format: int64
    get:
      tags:
        - Cadastro
      summary: Buscar aeronauta por Id
      operationId: buscarAeronautaPorId
      responses:
        200:
          description: Aeronauta encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aeronauta'
        404:
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    Fabricante:
      required:
        - nome
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: HELIBRÁS
    Aeronave:
      required:
        - marcas
        - apelido
        - modelo
        - numero_serie
        - categoria_registro
        - fabricante
      properties:
        id:
          type: integer
          format: int64
          example: 12
        apelido:
          type: string
          example: Águia
        marcas:
          type: string
          example: PP-PIT
        fabricante:
          $ref: '#/components/schemas/Fabricante'
        modelo:
          type: string
          example: AS350B2
        numero_serie:
          type: integer
          example: 4942
        horimetro:
          $ref: '#/components/schemas/Horimetro'
        categoria_registro:
          type: string
          example: SAE
    AeronaveRequest:
      required:
        - marcas
        - apelido
        - modelo
        - numero_serie
        - categoria_registro
        - fabricante_id
      properties:
        apelido:
          type: string
          example: Águia
        marcas:
          type: string
          example: PP-PIT
        fabricante_id:
          type: integer
          format: int64
          example: 1
        modelo:
          type: string
          example: AS350B2
        numero_serie:
          type: integer
          example: 4942
        categoria_registro:
          type: string
          example: SAE
    Error:
      properties:
        code:
          type: string
          example: 404
        message:
          type: string
          example: Valor inválido
      required:
        - code
        - message
    Horimetro:
      required:
        - hora_total_decimal
        - hora_manutencao_decimal
      properties:
        hora_total:
          type: string
          example: "04:48"
          readOnly: true
        hora_manutencao:
          type: string
          example: "03:12"
          readOnly: true
        hora_total_decimal:
          type: number
          format: double
          example: 4.8
        hora_manutencao_decimal:
          type: number
          format: double
          example: 3.2
    Horimetros:
      type: array
      items:
        $ref: '#/components/schemas/Horimetro'
    Pessoa:
      required:
        - cpf
        - nome
      properties:
        id:
          type: integer
          format: int64
          example: 1
        cpf:
          type: string
          example: 93226553054
        nome:
          type: string
          example: Rogerio Cabral
    Aeronauta:
      required:
        - pessoa
        - codigo_anac
      properties:
        id:
          type: integer
          format: int64
          example: 1
        pessoa:
          $ref: '#/components/schemas/Pessoa'
        codigo_anac:
          type: integer
          example: 10686
  requestBodies:
    FabricanteBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Fabricante'
    AeronaveBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AeronaveRequest'
    HorimetroBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Horimetro'
    PessoaBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Pessoa'
    AeronautaBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Aeronauta'
  responses:
    NotFound:
      description: Não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidParameter:
      description: Invalid parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'